Add updateLine variant of addOrUpdateLine
This will only update an existing PATTERN with a REPLACEMENT, but won't add anything should no match exist. Therefore, no --line LINE needs to be given, but both --update-match PATTERN and --replacement REPLACEMENT are mandatory. Because of that, --create-nonexisting does not make sense here, neither do --pre-line and --post-line.
The adaptations of addOrUpdateLine are surprisingly small; mostly it's the changed command-line argument parsing. The command is reused through a new wrapper script; integrating the functionality directly into addOrUpdateLine would have made the usage help really messy; in the wrapper we're going to massage that instead.
The functionality is enabled via the ADDORUPDATELINE_NO_APPEND environment variable; it makes the LINE never match (so triggering only happens through PATTERN), and turns off the isAppend flag for all passed files (instead of just for files other than the last).