Use cat to read the patch file so that patch --output=- error messages can be seen
patch is a bit unorthodox; it prints errors to stdout unless that is used for outputting the result, where it switches to stderr for errors.
The error message suppression on the shell input redirection would kill those, so switch to using cat (which will also generate "file does not exist" error message for us, so we save that check at the cost of slightly less readable error messages).