FIX: Correct status handling
For the default / --first logic, we can directly accept a success, mark a no-mod unless we've already had a success.
With --all, a success initializes to 0 and changes a previous 1 to 0, as an update is stronger than a no-mod. A no-mod initializes the status to 1, but will be overwritten by either success (to 0), or non-apply (to 4).