updateWithPatch: FIX: Actually pass patchArgs
We've only covered the patch arguments that require special handling in the script so far, and naturally for a pure test-driven development, the actual common patch arguments were forgotten.
Add tests for --ifdef and -p0 as two representatives, together with new test scaffolding for patch manipulation on the fly.
Minor: Pass --silent explicitly to doPatch().