containedOrUpdateWithPatch: Use --read-only=ignore --test-only
As that may be run without superuser rights (yet), as we're only checking so far. This change is mostly driven by checkedAddOrUpdate, as I don't want to add yet another special case for that. But I think it also make sense in general.
Separate parsing of --read-only into updateWithPatchReadOnlyArgs, and pass that only selectively on the actual processing, but not for --test-only and also not for obtaining the affected files.