updateWithPatch: ENH: Make --print-affected-files consider a passed --directory
Any relative patched files (and that's the purpose of using --directory) are output by patch relative to the directory. We could ask the client to do this translation (after all, it passed the --directory), but maybe the client also is just forwarding the arguments, and then we're already parsing and picking apart patch's arguments, anyway.
As there still might be absolute filespecs in the patch, we can't simply prepend the directory. Let's just change directory as well and let realpath translate to absolute paths; these are more reliable downstream (as an added bonus).