containedOrUpdate{Block,Line}: Message on status 1 also needs to mention no match
As these are missing the "AddOr" part, if the block doesn't exist yet, it will also fail (and not just add it), so exit status 1 cannot just be caused by contents that are already up-to-date, but also by the marker / pattern not being found in the file. The original message would be highly misleading in those cases.
Use the ADDORUPDATE{BLOCK,LINE}_NO_APPEND flag in the underlying implementations to add the additional verbiage into the message.