addOrUpdateWithSed: ENH: Allow special sed program for appending to the last file
In order to be consistent with addOrUpdateLine et al., the fallback of appending when no modification of an existing line is possible must only happen for the last file (unless --all is given).
As it's very hard to pass and evaluate a flag into the sed program itself, instead allow the insertion of dedicated sed program expressions only for the last file.